B follow-up: enforce Outcome invariant, move to common.model, fix spec + PgnCreate
Addresses three findings from the Phase B review:

[P2] Outcome compact constructor enforces the invariant the JavaDoc
already documented: winner is non-null iff termination is CHECKMATE.
Previously a caller could construct new Outcome(STALEMATE, Side.WHITE)
or new Outcome(CHECKMATE, null); now both throw IllegalArgumentException.

[P2] Outcome moved out of common.enums into common.model. Outcome is
a record/model; common.enums is for enums. Termination stays in
common.enums.

[P2] specification.md section 3.1 rewritten. It still described
checkmate / stalemate / insufficient material as automatic in the
"no further moves accepted by the move pipeline" sense, and it still
pointed at the deleted GameStatus enum and calculateGameStatus method.
The new text reflects the post-A1+A2+B posture:
  - termination is information, not enforcement
  - the table lists the five automatic terminations in
    calculateOutcome's python-chess precedence order (checkmate >
    IM > stalemate > 75-move > fivefold)
  - single-side IM is called out as diagnostic, not a termination
  - analyzer-driven dead positions remain opt-in
  - the API references point at Outcome and calculateOutcome

[P3] PgnCreate.calculateResultTagValue now derives the winner from
outcome.winner() rather than inverting board.getHavingMove(). With the
invariant in place this is straightforward and surfaces the structural
benefit of the new type at the actual consumer.

Test status: mvn test green (1133 run, 0 fail, 0 error, 3 skipped).
