+++ src/main/java/com/dlb/chess/common/utility/BasicChessUtility.java
@@ -6,7 +6,7 @@ import com.dlb.chess.board.Board;
 import com.dlb.chess.board.enums.Piece;
 import com.dlb.chess.board.enums.PieceType;
 import com.dlb.chess.board.enums.Side;
-import com.dlb.chess.common.enums.Outcome;
+import com.dlb.chess.common.model.Outcome;
 import com.dlb.chess.common.enums.Termination;
 import com.dlb.chess.model.LegalMove;
