+++ src/test/java/com/dlb/chess/test/unwinnability/oracle/ForcedLineOracle.java
@@ -4,7 +4,7 @@ import org.eclipse.jdt.annotation.Nullable;
 
 import com.dlb.chess.board.Board;
 import com.dlb.chess.board.enums.Side;
-import com.dlb.chess.common.enums.Outcome;
+import com.dlb.chess.common.model.Outcome;
 import com.dlb.chess.common.exceptions.ProgrammingMistakeException;
 import com.dlb.chess.common.utility.BasicChessUtility;
 import com.dlb.chess.common.utility.ListUtility;
